Add linting
ESLint is run on "npm test", and the configuration file is based on
eslint:recommended, eslint-config-node, and taking the file generated by
ESLint examining the files and tweaking it to fit the JS style we use in
other projects.

eslint-plugin-html is used to lint the inline <script> tags in the HTML.

Fixes any linter errors that eslint shows.

https://phabricator.endlessm.com/T20603