Fix size of app icon
The app icon was 256x257, with an extra blank row of pixels at the top.
This fixes it to be the required size of 256x256. Also moves it to an
icons/ directory and adjusts the location where the app looks for it,
and renames it to the app ID which will be required for Flathub.

https://phabricator.endlessm.com/T20603