Allow calling bin/hatch-previewer from other directories
Now non-flatpak standalone requires bin/hatch-previewer to be run from
the root of the repository.

Change path to the module to be the module path, not the current
directory, in the case where we are not in a flatpak package and not
imported as a module.

Otherwise if `PATH/TO/bin/hatch-previewer <path>` is called from an
ingester dir, the `index.js` of the ingester is executed!